wikimedia-kafka-python/wikimedia-kafka-python/requirements.txt
confluent-kafka
opensearch-py
typing-extensions
sseclient-py
